+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added more SQL translations including stringr and lubridate functions
 * Updated for compatibility with dplyr 0.8.3
 * Fixed errors when `na.rm = TRUE` is specified in some aggregate functions
+* Added `src_databases()` function (#36)
 * Made minor bugfixes and improvements
 
 # implyr 0.2.4